Add Flaky external compaction tests for cancellation
This refactors ExternalCompaction_2_IT so that the same tests can be run
using a Flaky Ample implementatino which will randomly return an UNKNOWN
status when submitting conditional mutations. The tests being run are
related to cancelling compaction jobs so this test makes sure the
rejection handlers related to job cancellation in CompactionCoordinator
work correctly, such as the rejection handler used inside of
CompactionCoordinator.compactionFailedForLevel

This implements one of the methods detected as missing code coverage
that is listed in #4710
